Fix cast issue into Item class
Divide int vars produces an integer result, so it was needed to cast
each variable to double type and return a double value.

+++ src/knapsack/Item.java
@@ -26,10 +26,10 @@ public class Item {
         this.weight = weight;
     }
     
-    public float getRealValue() {
-        float realValue;
+    public double getRealValue() {
+        double realValue;
         if(weight > 0) {
-            realValue = (float)(value / weight);
+            realValue = (double)value / (double)weight;
         } else {
             realValue = value;
         }
